// src/components/StoryModal/types.ts
export enum StroyTypes {
  Image = 'image',
  Video = 'video',
  Text = 'text',
}
// ...
export interface StoryViewProps {
  pause?: boolean;
  onVideoLoaded?: (arg0: OnLoadData) => void;
  videoDuration: number[];
  onVideoProgress?: (progressData?: OnProgressData) => void;
  onVideoEnd?: () => void;
  onImageLoaded?: () => void;
  stories: StoryType[];
  showSourceIndicator?: boolean;
  sourceIndicatorProps?: ActivityIndicatorProps;
  videoProps?: VideoProperties;
  index?: number;
  storyIndex?: number;
}
// ...
export type StoryType = {
  id: number;
  url?: string;
  thumbnail?: string;
  type?: string | 'image' | 'video' | 'text';
  duration?: number;
  isReadMore?: boolean;
  isViewed?: boolean;
  isPaused?: boolean;
  created?: string;
  storyId?: number;
  title?: string;
};